+++ src/main/java/com/app/services/CartService.java
@@ -10,7 +10,7 @@ public interface CartService {
 	
 	List<CartDTO> getAllCarts();
 	
-	CartDTO getCart(String emailId, Long cartId);
+	CartDTO getCart(String emailId, Integer cartId);
 	
 	CartDTO updateProductQuantityInCart(Long cartId, Long productId, Integer quantity);
